use stored procs for video data

+++ src/Maw.Domain/Rating.cs
@@ -2,7 +2,7 @@ namespace Maw.Domain
 {
     public class Rating
     {
-		public byte UserRating { get; set; }
+		public short UserRating { get; set; }
         public float AverageRating { get; set; }
     }
 }
